Performance testing using `tap benchmark test-plan --quiet --iterations 10000`

- main branch: Time spent per iteration: 5.16 ms (σ: 0.59 ms)
- With logging: Time spent per iteration: 3.85 ms (σ: 0.05 ms)
- Without logging: Time spent per iteration: 2.69 ms (σ: 0.06 ms)

Without logging: see commented out log messages in TestStep.cs

Main branch high variance seems to suggest that a significant improvement has been made with regards to generating garbage.